requirements.txt
flask
flask-wtf
wtforms
psycopg2-binary
bcrypt
email-validator
gunicorn
